+++ examples/data-sources/iosxr_banner/data-source.tf
@@ -1,0 +1,3 @@
+data "iosxr_banner" "example" {
+  banner_type = "login"
+}