+++ examples/resources/iosxr_banner/resource.tf
@@ -1,0 +1,4 @@
+resource "iosxr_banner" "example" {
+  banner_type = "login"
+  line        = " Hello user !"
+}