+++ examples/resources/iosxr_cdp/import.sh
@@ -1,0 +1,1 @@
+terraform import iosxr_cdp.example "Cisco-IOS-XR-um-cdp-cfg:/cdp"