+++ internal/provider/provider.go
@@ -299,7 +299,9 @@ func (p *iosxrProvider) Configure(ctx context.Context, req provider.ConfigureReq
 func (p *iosxrProvider) Resources(ctx context.Context) []func() resource.Resource {
 	return []func() resource.Resource{
 		NewGnmiResource,
+		NewBannerResource,
 		NewBGPASFormatResource,
+		NewCDPResource,
 		NewEVPNResource,
 		NewEVPNEVIResource,
 		NewEVPNGroupResource,
@@ -359,7 +361,9 @@ func (p *iosxrProvider) Resources(ctx context.Context) []func() resource.Resourc
 func (p *iosxrProvider) DataSources(ctx context.Context) []func() datasource.DataSource {
 	return []func() datasource.DataSource{
 		NewGnmiDataSource,
+		NewBannerDataSource,
 		NewBGPASFormatDataSource,
+		NewCDPDataSource,
 		NewEVPNDataSource,
 		NewEVPNEVIDataSource,
 		NewEVPNGroupDataSource,