+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 ## 0.1.10 (unreleased)
 
+- Add `iosxr_banner` resource and data source
+- Add `iosxr_cdp` resource and data source
 - Add `iosxr_extcommunity_opaque_set` resource and data source
 - Add `iosxr_segment_routing_v6` resource and data source
 - Add `iosxr_segment_routing_te` resource and data source